backend/pyproject.toml
[project]
name = "plainer-backend"
version = "0.1.0"
description = "Plainer - AI-powered collaborative workspace backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pyjwt>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "anthropic>=0.42.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-magic>=0.4.27",
    "redis>=5.2.0",
    "aiofiles>=24.1.0",
    "email-validator>=2.0.0",
    "aioboto3>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
